fix/improve nqueens
Unconditionally assigning to buf in filter was causing an error (wrong result value) on some runtimes after upgrading to Clang 21. This may be related to lazy evaluation of the range.

Moving the buf assignment into the transform fixes the problem, and also improves performance substantially - up to 10% on some runtimes.